// src/api/services/auth/models/omitCreatePlanFeatureDto.ts
/**
 * Generated by orval v7.0.1 🍺
 * Do not edit manually.
 * Growinvoice API
 * Enhance your business with Growinvoice API
 * OpenAPI spec version: 1.0
 */
import type { OmitCreatePlanFeatureDtoFeature } from "./omitCreatePlanFeatureDtoFeature";

export interface OmitCreatePlanFeatureDto {
	count: number;
	feature: OmitCreatePlanFeatureDtoFeature;
}